tests/wpt/tests/workers/WorkerLocation_hash_encoding.htm

<!DOCTYPE html>
<title> WorkerLocation.hash with url encoding string </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker("./support/WorkerLocation.js#question%3f");
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.hash, "#question%3f");
  });
});
</script>
